+++ src/QueueConfig.js
@@ -33,8 +33,7 @@ class QueueConfig {
       options = {},
       rpcTimeoutMs = 10000,
       rpcQueueMaxSize = 100,
-      logger = console,
-      exitOnConnectionClose = 2
+      logger = console
     } = config
 
     const rabbitMqOptions = new RabbitMqOptions(options)
@@ -43,7 +42,6 @@ class QueueConfig {
     this.rpcTimeoutMs = rpcTimeoutMs
     this.rpcQueueMaxSize = rpcQueueMaxSize
     this.logger = logger
-    this.exitOnConnectionClose = exitOnConnectionClose
   }
 }